+++ frontend/app/src/pages/superadmin/constants.ts
@@ -1,0 +1,2 @@
+export const pageSize = 20;
+export const visibleTabs = 7;
