+++ frontend/app/src/pages/superadmin/tableComponent/index.tsx
@@ -1,14 +1,14 @@
-import React, { useCallback, useEffect, useState } from 'react';
-import { useStores } from 'store';
+import React, { useState } from 'react';
 import moment from 'moment';
 import { EuiPopover, EuiText } from '@elastic/eui';
 import MaterialIcon from '@material/react-material-icon';
-import { BountyStatus, defaultBountyStatus } from '../../../store/main';
+import { BountyStatus } from '../../../store/main';
 import paginationarrow1 from '../header/icons/paginationarrow1.svg';
 import paginationarrow2 from '../header/icons/paginationarrow2.svg';
 import defaultPic from '../../../public/static/profile_avatar.svg';
 import copygray from '../header/icons/copygray.svg';
 import { dateFilterOptions, getBountyStatus } from '../utils';
+import { pageSize, visibleTabs } from '../constants.ts';
 import { colors } from './../../../config/colors';
 import { Bounty } from './interfaces.ts';
 
@@ -59,7 +59,11 @@ interface TableProps {
   paginatePrev?: () => void;
   paginateNext?: () => void;
   currentPage: number;
+  totalBounties: number;
+  paginationLimit: number;
   setCurrentPage?: React.Dispatch<React.SetStateAction<number>>;
+  activeTabs: number[];
+  setActiveTabs: React.Dispatch<React.SetStateAction<number[]>>;
 }
 
 interface ImageWithTextProps {
@@ -128,8 +132,6 @@ export const TextInColorBox = ({ status }: TextInColorBoxProps) => (
 
 export const MyTable = ({
   bounties,
-  startDate,
-  endDate,
   bountyStatus,
   setBountyStatus,
   dropdownValue,
@@ -138,19 +140,16 @@ export const MyTable = ({
   setDropdownValue,
   onChangeFilterByDate,
   currentPage,
-  setCurrentPage
+  setCurrentPage,
+  activeTabs,
+  setActiveTabs,
+  totalBounties,
+  paginationLimit
 }: TableProps) => {
-  const [totalBounties, setTotalBounties] = useState(0);
   const [isPopoverOpen, setIsPopoverOpen] = useState<boolean>(false);
   const onButtonClick = () => setIsPopoverOpen((isPopoverOpen: any) => !isPopoverOpen);
   const closePopover = () => setIsPopoverOpen(false);
-  const pageSize = 20;
-  const visibleTabs = 7;
 
-  const { main } = useStores();
-
-  const paginationLimit = Math.floor(totalBounties / pageSize) + 1;
-  const [activeTabs, setActiveTabs] = useState<number[]>([]);
 
   const updateBountyStatus = (e: any) => {
     if (bountyStatus && setBountyStatus && setDropdownValue) {
@@ -177,8 +176,6 @@ export const MyTable = ({
 
       dataNumber.push(nextPage);
       dataNumber.shift();
-      setActiveTabs(dataNumber);
-      console.log("Active Tabs ===", activeTabs, dataNumber)
     }
   };
 
@@ -201,36 +198,11 @@ export const MyTable = ({
     }
   };
 
-  const getTotalBounties = useCallback(async () => {
-    if (startDate && endDate) {
-      const totalBounties = await main.getBountiesCountByRange(String(startDate), String(endDate));
-      setTotalBounties(totalBounties);
-    }
-  }, [main, startDate, endDate]);
-
-  const getActiveTabs = useCallback(() => {
-    const dataNumber: number[] = [];
-    for (let i = 1; i <= Math.ceil(paginationLimit); i++) {
-      if (i > visibleTabs) break;
-      dataNumber.push(i);
-    }
-
-    setActiveTabs(dataNumber);
-  }, [paginationLimit]);
-
   const paginate = (page: number) => {
     if (setCurrentPage) {
       setCurrentPage(page);
     }
   }
-
-  useEffect(() => {
-    getTotalBounties();
-  }, [getTotalBounties]);
-
-  useEffect(() => {
-    getActiveTabs();
-  }, [getActiveTabs]);
 
   const color = colors['light'];
